multiple circuits gets checked first
If more than one circuit is connectd, it will color all the neopixels red. We also had to make sure that multiple circuits was at the top of the if statement and then followed by checking singular connections and then checks to see if nothing is connected. the order matters because if i have the check for a single circuit before multiple circuits, it will execute the code for the single code block and avoid the rest of the if else statements. Similar example of this is the FizzBuzz technical exam questions